requirements.txt
alabaster==0.7.9
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.0
argh==0.26.2
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.1
Babel==2.3.4
beautifulsoup4==4.5.3
billiard==3.3.0.23
boto==2.39.0
bumpversion==0.5.3
cached-property==1.3.0
celery==3.1.20
channels==1.0.2
click==6.6
constantly==15.1.0
contextlib2==0.5.4
coreapi==2.0.0
coverage==4.0
daphne==1.0.1
decorator==4.0.11
defusedxml==0.4.1
Django==1.9.2
django-allauth==0.26.1
django-autoslug==1.9.3
django-braces==1.8.1
django-compat==1.0.13
django-cors-headers==1.1.0
django-coverage-plugin==1.2.2
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-email-extras==0.3.3
django-environ==0.4.0
django-extensions==1.6.1
django-filter==1.0.1
django-floppyforms==1.6.2
django-hijack==2.1.2
django-hijack-admin==2.1.2
django-liked==1.0.2
django-mailgun==0.9.1
django-material==0.8.0
django-model-utils==2.4
django-modelcluster==2.0
django-multiselectfield==0.1.5
django-oauth-toolkit==0.10.0
django-orderable==4.0.4
django-post-office==2.0.8
django-push-notifications==1.4.1
django-redis==4.3.0
django-redis-sessions==0.5.6
django-rest-auth==0.8.2
django-rest-authemail==1.4.0
django-rest-swagger==2.0.5
django-safedelete==0.3.5
django-secure==1.0.1
django-simple-history==1.8.1
django-social-auth==0.7.28
django-sql-explorer==1.0
django-storages-redux==1.3.2
django-taggit==0.22.0
django-test-plus==1.0.11
django-treebeard==4.1.0
django-versatileimagefield==1.6.2
-e git+https://github.com/jrief/django-websocket-redis@31723a7ae2de62a88c2b0d17950611052730f244#egg=django_websocket_redis
django-websocket-request==0.1.1
django-widget-tweaks==1.4.1
djangorestframework==3.3.2
docutils==0.12
drfdocs==0.0.11
factory-boy==2.6.0
fake-factory==9999.9.9
flake8==2.4.1
future==0.15.2
gevent==1.1rc3
greenlet==0.4.10
gunicorn==19.4.5
hiredis==0.2.0
html5lib==0.999999999
httplib2==0.10.3
impactsdk==0.1.0
incremental==16.10.1
ipdb==0.8.1
ipython==5.2.2
ipython-genutils==0.1.0
isodate==0.5.4
itypes==1.1.0
Jinja2==2.8
jsonfield==1.0.3
kombu==3.0.35
lxml==3.6.4
MarkupSafe==0.23
mccabe==0.3.1
msgpack-python==0.4.8
nose==1.3.7
oauth2==1.9.0.post1
oauthlib==1.0.3
opbeat==3.4.0
openapi-codec==1.0.0
pathtools==0.1.2
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.1.1
pluggy==0.3.1
prompt-toolkit==1.0.13
psycopg2==2.6.2
ptyprocess==0.5.1
py==1.4.31
pyBarcode==0.7
pycrypto==2.6.1
pyflakes==0.8.1
Pygments==2.1.3
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.6.0
python-gnupg==0.4.0
python-magic==0.4.12
python-openid==2.2.5
python-resize-image==1.1.3
python3-openid==3.0.10
pytz==2015.7
PyYAML==3.11
qrcode==5.3
qrtools==0.0.2
raven==5.26.0
razorpay==0.2.0
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.7.0
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.1
sphinx-me==0.3
sphinx-rtd-theme==0.1.9
sqlparse==0.2.2
suds-jurko==0.6
tixdo-api-client==0.1.0
tmdbsdk==0.1.0
tox==2.1.1
traitlets==4.3.1
Twisted==16.6.0
txaio==2.6.0
ua-parser==0.7.1
unicode-slugify==0.1.3
unicodecsv==0.14.1
Unidecode==0.4.19
uritemplate==3.0.0
user-agents==1.0.1
virtualenv==15.1.0
vistasdk==0.1.0
wagtail==1.8.1
watchdog==0.8.3
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.11.3
whitenoise==2.0.6
Willow==0.4
xlwt==1.0.0
xmltodict==0.10.2
zeep==0.15.0
zope.interface==4.3.3

